requirements.txt
streamlit
pandas
numpy
plotly
pillow
scipy
hydralit_components
